Update build file and code review
