+++ .github/actions/gradle-setup/action.yml
@@ -3,9 +3,10 @@ description: "Setup Gradle"
 runs:
   using: "composite"
   steps:
-    # Install Java and cache MVD Gradle build.
-    - uses: actions/setup-java@v2
+    - uses: actions/setup-java@v3
       with:
         java-version: '17'
         distribution: 'temurin'
-        cache: 'gradle'
+
+    - name: Setup Gradle cache
+      uses: gradle/gradle-build-action@v2
