+++ extensions/common/trust-framework-policies/src/main/java/org/eclipse/edc/trustframework/policy/seeding/model/PolicyEntry.java
