+++ extensions/common/trust-framework-policies/src/main/resources/gaiax-policies-22_10.json
