+++ extensions/common/trust-framework-policies/src/test/resources/gaiax-vc.json
