--- extensions/trust-framework-policies/src/main/java/org/eclipse/edc/trustframework/policy/seeding/exception/PolicyEvaluationException.java
@@ -1,23 +1,0 @@
-/*
- *  Copyright (c) 2022 Amadeus
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Amadeus - initial API and implementation
- *
- */
-
-package org.eclipse.edc.trustframework.policy.seeding.exception;
-
-import org.eclipse.dataspaceconnector.spi.EdcException;
-
-public class PolicyEvaluationException extends EdcException {
-    public PolicyEvaluationException(String message) {
-        super(message);
-    }
-}
