+++ gradle.properties
@@ -1,10 +1,23 @@
-projectGroup=org.eclipse.dataspaceconnector.trustframework
-# defaultVersion is used when "-PidentityHubVersion...." is no supplied. Should always be equal to edcVersion!
+groupId=org.eclipse.dataspaceconnector.trustframework
+javaVersion=11
 defaultVersion=0.0.1-SNAPSHOT
+# for now, we're using the same version for the autodoc plugin, the processor and the runtime-metamodel lib, but that could
+# change in the future
+annotationProcessorVersion=0.0.1-20221104.2-SNAPSHOT
+edcGradlePluginsVersion=0.0.1-20221104.2-SNAPSHOT
+metaModelVersion=0.0.1-20221104.2-SNAPSHOT
+#
 edcGroup=org.eclipse.dataspaceconnector
-edcVersion=0.0.1-20221011-SNAPSHOT
-edcGradlePluginsVersion=0.0.1-SNAPSHOT
-assertj=3.23.1
+edcVersion=0.0.1-20221104-SNAPSHOT
+# information required for publishing artifacts:
+edcDeveloperId=mspiekermann
+edcDeveloperName=Markus Spiekermann
+edcDeveloperEmail=markus.spiekermann@isst.fraunhofer.de
+edcScmConnection=scm:git:git@github.com:eclipse-dataspaceconnector/TrustFrameworkAdoption.git
+edcWebsiteUrl=https://github.com/eclipse-dataspaceconnector/TrustFrameworkAdoption.git
+edcScmUrl=https://github.com/eclipse-dataspaceconnector/TrustFrameworkAdoption.git
+#
 jupiterVersion=5.9.0
 mockitoVersion=4.8.0
+assertj=3.23.1
 jsonpathVersion=2.7.0
