+++ spi/common/trust-framework-spi/build.gradle.kts
