+++ spi/common/trust-framework-spi/src/main/java/org/eclipse/edc/trustframework/credentials/spi/VerifiableCredential.java
