+++ spi/common/trust-framework-spi/src/test/java/org/eclipse/edc/trustframework/credentials/spi/VerifiableCredentialTest.java
